+++ frontend/src/bundles/templates-page/store/actions.ts
@@ -6,7 +6,7 @@ import { type AsyncThunkConfig } from '~/bundles/common/types/types.js';
 import {
     type TemplateUpdateItemRequestDto,
     type TemplateUpdateItemResponseDto,
-} from '../types/types.js';
+} from '../../edit-temlate/types/types.js';
 import { name as sliceName } from './slice.js';
 
 type EditTemplayePayload = Partial<TemplateUpdateItemRequestDto> & {
