+++ frontend/src/bundles/templates-page/store/index.ts
