+++ frontend/src/bundles/templates-page/templates.ts
