+++ frontend/src/framework/store/store.package.ts
@@ -8,12 +8,12 @@ import { configureStore } from '@reduxjs/toolkit';
 import { authApi } from '~/bundles/auth/auth.js';
 import { reducer as authReducer } from '~/bundles/auth/store/';
 import { AppEnvironment } from '~/bundles/common/enums/enums.js';
-import { reducer as templatesReducer } from '~/bundles/edit-temlate/store/';
-import { templateApi } from '~/bundles/edit-temlate/templates.js';
 import { paymentApi } from '~/bundles/payment/payment.js';
 import { reducer as paymentReducer } from '~/bundles/payment/store';
 import { profileApi } from '~/bundles/profile/profile';
 import { reducer as profileReducer } from '~/bundles/profile/store/';
+import { reducer as templatesReducer } from '~/bundles/templates-page/store';
+import { templateApi } from '~/bundles/templates-page/templates.js';
 import { reducer as usersReducer } from '~/bundles/users/store/';
 import { userApi } from '~/bundles/users/users.js';
 import { type IConfig } from '~/framework/config/config.js';
